Update the project

+++ layered-architecture/subprojects/domain/src/main/java/com/k1m743hyun/data/dto/request/DemoRequestDto.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class DemoRequestDto {
-    private int id;
+
+    private Long id;
     private String name;
 }
