+++ layered-architecture/subprojects/domain/src/main/java/com/k1m743hyun/repository/DemoRepository.java
@@ -4,7 +4,8 @@ import com.k1m743hyun.data.entity.Demo;
 import java.util.Optional;
 
 public interface DemoRepository {
-    Optional<Demo> findById(int id);
+
+    Optional<Demo> findById(Long id);
 
     Optional<Demo> findByName(String name);
 }
